Fix more unsigned int / int cases that should be NS[U]Integer
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@31355 72102866-910b-0410-8b05-ffd578937521

+++ Source/GSAttributedString.m
@@ -71,7 +71,7 @@ static	NSDictionary	*blank;
 - (id) initWithString: (NSString*)aString
 	   attributes: (NSDictionary*)attributes;
 - (NSString*) string;
-- (NSDictionary*) attributesAtIndex: (unsigned)index
+- (NSDictionary*) attributesAtIndex: (NSUInteger)index
 		     effectiveRange: (NSRange*)aRange;
 
 @end
@@ -86,7 +86,7 @@ static	NSDictionary	*blank;
 - (id) initWithString: (NSString*)aString
 	   attributes: (NSDictionary*)attributes;
 - (NSString*) string;
-- (NSDictionary*) attributesAtIndex: (unsigned)index
+- (NSDictionary*) attributesAtIndex: (NSUInteger)index
 		     effectiveRange: (NSRange*)aRange;
 - (void) setAttributes: (NSDictionary*) attributes
 		 range: (NSRange)range;
@@ -496,7 +496,7 @@ _attributesAtIndexEffectiveRange(
   return AUTORELEASE([_textChars copyWithZone: NSDefaultMallocZone()]);
 }
 
-- (NSDictionary*) attributesAtIndex: (unsigned)index
+- (NSDictionary*) attributesAtIndex: (NSUInteger)index
 		     effectiveRange: (NSRange*)aRange
 {
   return _attributesAtIndexEffectiveRange(
@@ -603,7 +603,7 @@ SANITY();
   return _textProxy;
 }
 
-- (NSDictionary*) attributesAtIndex: (unsigned)index
+- (NSDictionary*) attributesAtIndex: (NSUInteger)index
 		     effectiveRange: (NSRange*)aRange
 {
   unsigned	dummy;
